Data fecthing refactor
