+++ multi-manager-dashboard/src/utils/calculateStrategyAPR.ts
